adding view model for the lunar phase page
