+++ src/react/lunarPhasePage/lunarPhasePage.tsx
@@ -1,42 +1,8 @@
-import { useEffect, useState } from "react";
 import { TableDisplay } from "../table/tableDisplay";
+import { useLunarPhasePageViewModel } from "./lunarPhasePageViewModel";
 
 const LunarPhasePage = () => {
-  const [lunarCount, setLunarCount] = useState([]);
-  const [animalList, setAnimalList] = useState([]);
-
-  const adjustLunarPhaseCount = (list: {moonPhase: string, animalName: string, callCount: number}[]):void => {
-    const newList = [];
-    var currentObj = {};
-    list.forEach((animalCall: {moonPhase: string, animalName: string, callCount: number}) => {
-      if (currentObj?.moonPhase !==  animalCall.moonPhase) {
-        if(Object.keys(currentObj).length > 0) {newList.push(currentObj)}
-        currentObj = { moonPhase: animalCall.moonPhase, total: 0};
-      }
-      const name = animalCall.animalName.replaceAll(" ", "");
-      currentObj[name] = animalCall.callCount;
-      currentObj.total += animalCall.callCount;
-      
-    });
-    newList.push(currentObj);
-    setLunarCount(newList)
-  }
-
-  const fetchLunarPhaseCount = (): void => {
-    const lunarList = window?.electron?.sendSync('get-lunarPhaseCount');
-    adjustLunarPhaseCount(lunarList);
-  }
-
-  const fetchAnimalSpecies = ():void => {
-    const animalList =  window?.electron?.sendSync('get-animalSpecies');
-    setAnimalList(animalList);
-  }
-
-
-  useEffect(() => {
-    fetchLunarPhaseCount();
-    fetchAnimalSpecies();
-  }, []);
+  const {animalList, lunarCount} = useLunarPhasePageViewModel();
 
   return <>
     <h1>Lunar Phase Page</h1>
